fix(wev-31): refactor legislative Sidebar
